Remove Dependancy on scp and hex editor, add Debug Bit
Instead of using a hex editor and transferring files using SCP , use the router's built in tools (dd,wc,grep and hexdump) to do the same. Also make sure that the user does a thorough sanity check on the bin file to avoid bricks by both checking the before and after hex states and the file size as seen on the adslgr forum as common failures by users that led to bricking of the devices. Additionally add the debug bit modification as it gives full access to bootloader in order to update it for even further debranding by upgrading with ATUB using the bootloader provided in the GPL sources since I suspect that is the only variable that can additionally stop debranding, however updating the bootloader requires tftp and a serial connection and thus it is not mentioned in this guide but this commit should be indexed by search engines for future reference of other people trying to debrand zyxel related routers. macadress was also corrected to macaddress